+++ hexagonal/src/main/java/com/example/hexagonal/application/core/usecase/UpdateCustomerUseCase.java
@@ -1,0 +1,32 @@
+package com.example.hexagonal.application.core.usecase;
+
+import com.example.hexagonal.application.core.domain.Customer;
+import com.example.hexagonal.application.ports.in.FindCustomerByIdInputPort;
+import com.example.hexagonal.application.ports.out.FindAddressByZipCodeOutputPort;
+import com.example.hexagonal.application.ports.out.UpdateCustomerOutputPort;
+
+public class UpdateCustomerUseCase {
+
+    private final FindCustomerByIdInputPort findCustomerByIdInputPort;
+
+    private final FindAddressByZipCodeOutputPort findAddressByZipCodeOutputPort;
+
+    private final UpdateCustomerOutputPort updateCustomerOutputPort;
+    
+    public UpdateCustomerUseCase(
+        FindCustomerByIdInputPort findCustomerByIdInputPort,
+        FindAddressByZipCodeOutputPort findAddressByZipCodeOutputPort,
+        UpdateCustomerOutputPort updateCustomerOutputPort
+        ) {
+        this.findCustomerByIdInputPort = findCustomerByIdInputPort;
+        this.findAddressByZipCodeOutputPort = findAddressByZipCodeOutputPort;
+        this.updateCustomerOutputPort = updateCustomerOutputPort;
+    }
+
+    public void update(Customer customer, String zipCode) {
+        findCustomerByIdInputPort.findById(customer.getId());
+        var address = findAddressByZipCodeOutputPort.find(zipCode);
+        customer.setAddress(address);
+        updateCustomerOutputPort.update(customer);
+    }
+}