+++ hexagonal/src/main/java/com/example/hexagonal/application/ports/out/DeleteCustomerByIdOutputPort.java
@@ -1,0 +1,6 @@
+package com.example.hexagonal.application.ports.out;
+
+public interface DeleteCustomerByIdOutputPort {
+
+    void delete(String id);
+}