+++ CHANGELOG.md
@@ -12,3 +12,11 @@
     - Creating UseCase.
     - Creating Adapter.
     - Creating Controller.
+- CRUD - Update
+    - Creating UseCase.
+    - Creating Adapter.
+    - Creating Controller.
+- CRUD - Delete
+    - Creating UseCase.
+    - Creating Adapter.
+    - Creating Controller.